+++ EUIJIN.md
@@ -1,0 +1,56 @@
+## 좋았던 점
+3주만에 이정도로 짠게 매우매우 신기하네요!
+역시 주함 리더 최고...
+
+또한, 4명이 같이 했음에도 일관성이 어느정도 잘 지켜진 모습이 너무 대단해요!
+
+좜이가 전체적인 세팅을 잘 해놓아서 그런 것 같아요.
+
+특히 API 계층을 분리해놓은 것이 인상깊었습니다!
+
+또한 컴포넌트가 비대해지지 않게 잘 나눈 것 같아요.
+읽기에 크게 불편함이 없었답니다~~
+
+그리고 웹 접근성 진짜 굿이네욤 ㅎㅎ
+멋진 프론트엔드 개발자!!
+
+## 아쉬운 점
+### PATH 상수화
+라우팅은 자주 활용되므로, 변경사항에 예방하기 위해 상수화하는 것이 좋겠습니당
+
+### 변수명
+Ic, St, CTA, Desc 등 줄임말이 많은데요.
+제 개인적인 비선호일 수도 있다만!!, 길지 않는 한 잘 줄이지 않습니당.
+
+### 상수 선언
+상수 선언 위치가 컴포넌트 내부인 경우가 많아요!
+렌더링 최적화 관점에서 많이 생각해보면 좋을것같아요.
+컴포넌트는 함수이고, 상태변화에 따라 매번 새로 그려지는데, 그때마다 새로 선언되어야할 변수인지 아닌지
+판단하여 컴포넌트 밖에 두거나, 상수 파일로 분리하거나 할 수 있을 것 같아요.
+
+## 
+
+## 개선하면 좋을 점
+### Portal을 이용한 Modal 정의
+
+### Suspense 사용
+react-query나 swr을 사용할 때의 장점중 하나가 suspense를 쉽게 사용할 수 있다는 거에요.
+데이터 패칭 라이브러리 없이는 suspense를 쓰기 많이 까다롭답니다.
+왜 데이터 패칭 라이브러리에서는 suspense가 동작하는지 공부해보는 것도 많이 도움이 될 것 같아요.
+조금 말씀을 드리면
+suspense는 dom이 그려지기 전에 발생하는 데이터 통신에서 로딩 화면을 띄워주는 기능입니다.
+
+데이터 패칭 라이브러리 없이는 useEffect내에서 get 통신을 해옵니다. 
+이는 DOM이 mount된 후에 통신을 시작한다는 것입니다.
+
+데이터패칭 라이브러리는 dom mount 이전부터 데이터 통신을 가능하게 합니다.
+이 원리는... 좀 어렵더군요 허허.
+
+### Context API로만 할수는 없을까?
+요새 Context API + react-query 조합이 유행인 걸로 알고 있습니다!
+Recoil은 Context API에 비해 무엇이 좋고,
+어떤 경우에 Context API만 써도 충분한지 생각해보면 좋을 것 같아요
+
+## 타입 재활용
+모두 보진 못했지만, 비슷한 타입이 여러 곳에서 선언되어있다는 느낌을 받았어요.
+Pick, Omit, Partial 등의 유틸리티 타입을 이용해 하나하나 정의해보면 좋겠습니다.
