+++ src/App.tsx
@@ -6,14 +6,16 @@ import { activeState, activeStateModal } from "./core/atom/menuBar";
 import Router from "./Router";
 import { ModalStyle } from "./style/modalStyle";
 
+// 이 함수처럼, 컴포넌트 내에 속할 필요가 없는 녀석이 있음.
+// 컴포넌트 바깥에 선언, 컴포넌트 안에 선언 각각 어떤 의미를 지니는지 생각해보면 좋을것 같아욤! (렌더링 관점에서~~!)
+function setScreenSize() {
+  const vh = window.innerHeight * 0.01;
+  document.documentElement.style.setProperty("--vh", `${vh}px`);
+}
+
 export default function App() {
   const isActive = useRecoilValue(activeState);
   const isActiveModal = useRecoilValue(activeStateModal);
-
-  function setScreenSize() {
-    const vh = window.innerHeight * 0.01;
-    document.documentElement.style.setProperty("--vh", `${vh}px`);
-  }
 
   useEffect(() => {
     setScreenSize();
