+++ src/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { RecoilRoot } from "recoil";
@@ -6,7 +7,6 @@ import { ThemeProvider } from "styled-components";
 import App from "./App";
 import { GlobalStyle } from "./style/globalStyle";
 import theme from "./style/theme";
-
 const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
 root.render(
   <React.StrictMode>
